requirements.txt
python-dotenv~=1.0.1
fastapi~=0.110.2
langchain~=0.2.16

pydantic==1.10.18
langchain-community==0.2.17
langchain-core==0.2.40
langgraph==0.2.21
fastapi==0.115.0
uvicorn==0.30.2
